UP-2832 Don't remove guest layout from cache
git-svn-id: https://source.jasig.org/uPortal/trunk@21849 f5dbab47-78f9-eb45-b975-e544023573eb

+++ uportal-war/src/main/java/org/jasig/portal/layout/dlm/DistributedLayoutManager.java
@@ -156,8 +156,10 @@ public class DistributedLayoutManager implements IUserLayoutManager, IFolderLoca
 
     @Override
     public void afterPropertiesSet() throws Exception {
-        // Ensure a new layout gets loaded whenever a user logs in...
-        this.layoutCachingService.removeCachedLayout(owner, profile);
+        // Ensure a new layout gets loaded whenever a user logs in exceot for guest users
+        if (!owner.isGuest()) {
+            this.layoutCachingService.removeCachedLayout(owner, profile);
+        }
         
         this.loadUserLayout();
